+++ libraries/Sleep.h
@@ -9,17 +9,17 @@ extern "C" {
 namespace Lifeboxes  {
   const double SECONDS = 10 * micros();
   const double HOURS = 60 * 60 * SECONDS;
-  const double DEFAULT_SLEEPING_INTERVAL = ESP.deepSleepMax();
+  const uint64_t DEFAULT_SLEEPING_INTERVAL = ESP.deepSleepMax();
 
   class Sleep {
     public:
-      Sleep(uint32_t cyclesToSleep, uint32_t sleepingInterval = DEFAULT_SLEEPING_INTERVAL);
+      Sleep(uint32_t cyclesToSleep, uint64_t sleepingInterval = DEFAULT_SLEEPING_INTERVAL);
       bool isTimeToWakeUp();
       void goToSleep();
     private:
       uint32_t _cyclesSlept;
       uint32_t _cyclesToSleep;
-      uint32_t _sleepingInterval;
+      uint64_t _sleepingInterval;
       bool _wasResetFromSleepUp();
       void _updateCyclesSlept();
       void _readCyclesSlept();
